bugs fixes

+++ src/tui.h
@@ -18,7 +18,7 @@ typedef struct TUI_InputForm {
 
     size_t size;
     size_t cap;
-    size_t selected_option;
+    int    selected_option;
 
     size_t str_len;
 } TUI_InputForm;
